Extract LockAndTimestampServices out from TransactionManagers
