+++ atlasdb-dagger/src/main/java/com/palantir/atlasdb/services/LockAndTimestampModule.java
@@ -17,6 +17,7 @@ package com.palantir.atlasdb.services;
 
 import javax.inject.Singleton;
 
+import com.palantir.atlasdb.factory.LockAndTimestampServices;
 import com.palantir.atlasdb.factory.TransactionManagers;
 import com.palantir.lock.RemoteLockService;
 import com.palantir.lock.impl.LockServiceImpl;
@@ -30,7 +31,7 @@ public class LockAndTimestampModule {
 
     @Provides
     @Singleton
-    public TransactionManagers.LockAndTimestampServices provideLockAndTimestampServices(ServicesConfig config) {
+    public LockAndTimestampServices provideLockAndTimestampServices(ServicesConfig config) {
         return TransactionManagers.createLockAndTimestampServices(
                 config.atlasDbConfig(),
                 resource -> { },
@@ -40,13 +41,13 @@ public class LockAndTimestampModule {
 
     @Provides
     @Singleton
-    public TimestampService provideTimestampService(TransactionManagers.LockAndTimestampServices lts) {
+    public TimestampService provideTimestampService(LockAndTimestampServices lts) {
         return lts.time();
     }
 
     @Provides
     @Singleton
-    public RemoteLockService provideLockService(TransactionManagers.LockAndTimestampServices lts) {
+    public RemoteLockService provideLockService(LockAndTimestampServices lts) {
         return lts.lock();
     }
