+++ atlasdb-server/src/main/java/com/palantir/atlasdb/server/AtlasDbServer.java
@@ -24,7 +24,7 @@ import com.google.common.collect.ImmutableMap;
 import com.palantir.atlasdb.factory.ImmutableLockAndTimestampServices;
 import com.palantir.atlasdb.factory.Leaders;
 import com.palantir.atlasdb.factory.ServiceDiscoveringAtlasSupplier;
-import com.palantir.atlasdb.factory.TransactionManagers.LockAndTimestampServices;
+import com.palantir.atlasdb.factory.LockAndTimestampServices;
 import com.palantir.atlasdb.server.config.AtlasDbServerConfiguration;
 import com.palantir.atlasdb.server.config.ClientConfig;
 import com.palantir.leader.LeaderElectionService;
